dadw

+++ app/game-admin/dashboard/about-us/AboutItem.jsx
@@ -21,27 +21,35 @@ const AboutItem = ({ id, heading, content }) => {
   const handleDelete = async (e, id) => {
     e.stopPropagation();
 
-    setLoader(true);
-
-    const res = await axios.delete(
-      `${process.env.NEXT_PUBLIC_BASE_URL}/api/admin/aboutus`,
-
-      {
-        headers: { id: id, "Cache-Control": "no-store" },
-      }
+    const confirmDelete = window.confirm(
+      "Are you sure you want to delete this About-Us Topic?"
     );
 
-    if (res.data.status !== true) {
+    if (confirmDelete) {
+      setLoader(true);
+
+      const res = await axios.delete(
+        `${process.env.NEXT_PUBLIC_BASE_URL}/api/admin/aboutus`,
+
+        {
+          headers: { id: id, "Cache-Control": "no-store" },
+        }
+      );
+
+      if (res.data.status !== true) {
+        setLoader(false);
+        return toast.error(res.data.message);
+      }
+
+      toast.success(res.data.message);
+
+      router.push("/game-admin/dashboard/about-us", { scroll: false });
+      router.refresh();
+
       setLoader(false);
-      return toast.error(res.data.message);
+    } else {
+      return toast.info("Item not deleted");
     }
-
-    toast.success(res.data.message);
-
-    router.push("/game-admin/dashboard/about-us", { scroll: false });
-    router.refresh();
-
-    setLoader(false);
   };
 
   return (
